requirements.txt
pytesseract
PyAutoGUI
pillow